refactoring

+++ keyword/keyword.go
@@ -2,5 +2,5 @@ package keyword
 
 type Keyword struct {
 	Word string
-	Rate int16
+	Rate int
 }
